drain changes
Removed home and set the target on expected balls to the trough. That
was the problem before. the default target was “playfield” so when it
found a ball in the drain it setup the eject chain to send a ball to
the pf, which caused it to kick one out of the trough first.