Fixed World Tour Achievements, tweaks to multiball, added 2ball mb to WT
jukebox_2_ball is now LONGHORN MULTIBALL, and is used for the end of
the Jukebox mode and the World Tour mode. We can pull it up at any
time, actually with the start_jukebox_two_ball event. I kept the mode
name the same so I didn’t have to refactor anything.